Fixed paths visibility
