+++ base_miner/predict.py
@@ -2,8 +2,7 @@
 # Copyright © 2023 Foundry Digital
 
 # Import modules that already exist or can be installed using pip
-from datetime import datetime, timedelta
-from threading import Thread
+from datetime import datetime
 from base_miner.model import retrain_and_save
 import json
 import joblib
@@ -14,7 +13,8 @@ from sklearn.preprocessing import MinMaxScaler
 # from base_miner.model import create_and_save_base_model_lstm, create_and_save_base_model_regression
 
 # import custom defined files
-from base_miner.get_data import load_df, save_df, merge_dfs, prep_data, scale_data, round_down_time
+from base_miner.get_data import prep_data, scale_data, round_down_time
+from neurons.utils import save_df
 
 
 def save_model_retraining_args(X_scaled: np.ndarray, y_scaled: np.ndarray):
@@ -64,7 +64,7 @@ def predict(timestamp: str, model, scaler: MinMaxScaler or None = None, X_scaled
             method="ffill")
 
     data.fillna(0, inplace=True)
-    save_df(data)
+    save_df(data, "../mining_models/GSPC.csv")
 
     # The timestamp sent by the validator need not be associated with an exact 5m interval
     # It's on the miners to ensure that the time is rounded down to the last completed 5 min candle
@@ -96,34 +96,35 @@ def predict(timestamp: str, model, scaler: MinMaxScaler or None = None, X_scaled
 
 # Uncomment this section if you wanna do a local test without having to run the miner
 # on a subnet. This main block (kinda) mimics the actual validator response being sent
-# if (__name__ == '__main__'):
-#     #     import yfinance as yf
-#     import pickle
-#
-#     #     import pmdarima as pm
-#     #
-#     #     # data = prep_data(False)
-#     #     # data = yf.download('^GSPC', period='60d', interval='5m')
-#     #     # data.reset_index(inplace=True)
-#     #     # scaler, X, y = scale_data(data)
-#     #     # print(data.tail())
-#     #     # # # # mse = create_and_save_base_model_regression(scaler, X, y)
-#     #     # # # #
-#     #     # # # # model = joblib.load('mining_models/base_linear_regression.joblib')
-#     #     # # # #
-#     #     # # # ny_timezone = timezone('America/New_York')
-#     #     # # # current_time_ny = datetime.now(ny_timezone) + timedelta(days=-1)  # for testing purposes
-#     #     # # # timestamp = current_time_ny.isoformat()
-#     timestamp = "2024-06-24T15:55:29.139514-04:00"
-#     #     # # # #
-#     #     # # # from base_miner.model import create_and_save_base_model_lstm
-#     #     # # #
-#     with open("mining_models/arimax_model.pkl", "rb") as model_f:
-#         model = pickle.load(model_f)
-#     #     # model = load_model("mining_models/base_lstm_new.h5")
-#     #     # model = retrain_and_save(X, y, "mining_models/base_lstm_new.h5")
-#     prediction = predict(timestamp, model=model, type='arimax')
-#     print(prediction)
+if (__name__ == '__main__'):
+    #     import yfinance as yf
+    from tensorflow.keras.models import load_model
+    import pickle
+
+    #     import pmdarima as pm
+    #
+    #     # data = prep_data(False)
+    #     # data = yf.download('^GSPC', period='60d', interval='5m')
+    #     # data.reset_index(inplace=True)
+    #     # scaler, X, y = scale_data(data)
+    #     # print(data.tail())
+    #     # # # # mse = create_and_save_base_model_regression(scaler, X, y)
+    #     # # # #
+    #     # # # # model = joblib.load('mining_models/base_linear_regression.joblib')
+    #     # # # #
+    #     # # # ny_timezone = timezone('America/New_York')
+    #     # # # current_time_ny = datetime.now(ny_timezone) + timedelta(days=-1)  # for testing purposes
+    #     # # # timestamp = current_time_ny.isoformat()
+    timestamp = "2024-06-26T15:55:29.139514-04:00"
+    #     # # # #
+    #     # # # from base_miner.model import create_and_save_base_model_lstm
+    #     # # #
+    with open("../mining_models/arimax_model.pkl", "rb") as model_f:
+        model = pickle.load(model_f)
+    #     # model = load_model("mining_models/base_lstm_new.h5")
+    #     # model = retrain_and_save(X, y, "mining_models/base_lstm_new.h5")
+    prediction = predict(timestamp, model=model, type='arimax')
+    print(prediction)
 
 # df = pd.read_csv("GSPC.csv").set_index("Datetime")
 # print(df.head())
