+++ neurons/miner.py
@@ -17,16 +17,13 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 # DEALINGS IN THE SOFTWARE.
 
-import argparse
 import time
 import typing
 import pickle
 import bittensor as bt
 
-from base_miner.model import retrain_and_save
 from base_miner.predict import predict
-from base_miner.get_data import prep_data, scale_data
-from base_miner.scheduler import main as schedule
+from neurons.scheduler import main as schedule
 
 # import predictionnet
 # Bittensor Miner Template:
@@ -38,7 +35,6 @@ from predictionnet.base.miner import BaseMinerNeuron
 # ML imports
 # import tensorflow
 # import numpy as np
-from tensorflow.keras.models import load_model
 from huggingface_hub import hf_hub_download
 
 import os
